+++ DesignToolbox/DesignToolbox (Light)/Views/Components/Dialogs/DialogsView.swift
@@ -18,7 +18,10 @@ struct DialogsView: View {
     var body: some View {
         NavigationView {
             List {
+                #if os(tvOS)
                 NavigationLink("Alert Message", destination: AlertMessageView())
+                #endif
+                NavigationLink("Inline Alert", destination: InlineAlertMessageView())
             }
         }
     }