+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Library] Signatures of control-item-based components (Orange-OpenSource/ouds-ios#1314)
 
+- [Library] `inline alert` component (Orange-OpenSource/ouds-ios#1307)
+- [Library] `alert message` component (Orange-OpenSource/ouds-ios#1159)
+
 ### Fixed
 
 - [Library] Size of loader for `button` component (Orange-OpenSource/ouds-ios#1296)